functions/prompt_injection_scanner/requirements.txt
onnxruntime>=1.17.0,<2.0
transformers>=4.40.0,<5.0
huggingface_hub>=0.23.0,<1.0
PyPDF2>=3.0.0,<4.0
python-docx>=1.1.0,<2.0
python-pptx>=0.6.23,<1.0
boto3>=1.34.0,<2.0
